tts_with_emotions/tts_with_emotions/requirements.txt

std_msgs
scipy
torch
transformers
suno-bark
